+++ Editor/src/Viewports/Private/Viewport.cpp
@@ -80,7 +80,7 @@ namespace AGE
 					EditorLayer::Get().SetViewportSize(Vector2(View[0], View[1]));
 					SpecView = View;
 				}
-				ImGui::Image((ImTextureID)(uintptr_t)FrameBuffer->GetColorAttachmentRendererID(Index), ImVec2(SpecView[0], SpecView[1]), ImVec2(0, 1), ImVec2(1, 0));
+				ImGui::Image((ImTextureID)(uintptr_t)FrameBuffer->GetColorAttachmentRendererID((uint32_t)Index), ImVec2(SpecView[0], SpecView[1]), ImVec2(0, 1), ImVec2(1, 0));
 
 				if (ImGui::BeginDragDropTarget())
 				{
